relax delete-after-X failed attempts logic
We need to account for cases like captive portals, which may trigger
false positives with current logic, forcing UA to unregister
report-uri's where it shouldn't.

We need to experiment with exact parameters on how to best handle this,
hence intentionally changing the spec language to be more open. Once/if
we converge on a set of rules, we can update the spec accordingly.