mark server errors (5xx) as optional
Strictly speaking, 5xx is out of scope for "network error" as a 5xx
constitutes a valid request-response exchange. However, server errors
are often encountered due to misbehaving proxy/gateway services which
may be outside of control of the operating site (e.g. site is being
filtered or blocked) and we want to allow the user agent to notify the
site when such cases are encountered.