Remove redundant URL
The Reporting API now lets us provide a custom URL to associate the
report with.  The default is to use the containing document's URL,
which isn't relevant for a NEL report.  This lets us remove the `uri`
field from the NEL body, removing a source of confusion about why
they're different.