Remove redundant URL (#90)
The Reporting API now lets us provide a custom URL to associate the
report with.  The default is to use the containing document's URL,
which isn't relevant for a NEL report; for NEL reports we now use the
URL of the original request as the URL of the report.  This was previously
stored in the NEL body's `uri` field; this is now redundant, and can be
removed.